Project-T1-MedPack-PrimaryAgent-MultiAgent-Qdrant-FastAPI/requirements.txt
langchain
langchain_google_genai
langgraph
pydantic
fastapi
uvicorn
python-multipart
Pillow
python-dotenv
pandas
numpy
qdrant-client
Scikit-learn
pydantic-settings
ipykernel 